src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o /Users/ryan/src/AzureHayMaker/src/requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.2
    # via
    #   azure-functions-durable
    #   microsoft-kiota-authentication-azure
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anthropic==0.73.0
    # via azure-haymaker (pyproject.toml)
anyio==4.11.0
    # via
    #   anthropic
    #   httpx
attrs==25.4.0
    # via aiohttp
azure-common==1.1.28
    # via
    #   azure-mgmt-authorization
    #   azure-mgmt-containerinstance
    #   azure-mgmt-resource
    #   azure-mgmt-resourcegraph
    #   azure-mgmt-subscription
azure-core==1.36.0
    # via
    #   azure-cosmos
    #   azure-data-tables
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-mgmt-core
    #   azure-monitor-ingestion
    #   azure-servicebus
    #   azure-storage-blob
    #   microsoft-kiota-authentication-azure
    #   msrest
azure-cosmos==4.14.2
    # via azure-haymaker (pyproject.toml)
azure-data-tables==12.7.0
    # via azure-haymaker (pyproject.toml)
azure-functions==1.24.0
    # via
    #   azure-haymaker (pyproject.toml)
    #   azure-functions-durable
azure-functions-durable==1.4.0
    # via azure-haymaker (pyproject.toml)
azure-identity==1.25.1
    # via
    #   azure-haymaker (pyproject.toml)
    #   msgraph-sdk
azure-keyvault-secrets==4.10.0
    # via azure-haymaker (pyproject.toml)
azure-mgmt-authorization==4.0.0
    # via azure-haymaker (pyproject.toml)
azure-mgmt-containerinstance==10.1.0
    # via azure-haymaker (pyproject.toml)
azure-mgmt-core==1.6.0
    # via
    #   azure-mgmt-authorization
    #   azure-mgmt-containerinstance
    #   azure-mgmt-resource
    #   azure-mgmt-resourcegraph
    #   azure-mgmt-subscription
azure-mgmt-resource==24.0.0
    # via azure-haymaker (pyproject.toml)
azure-mgmt-resourcegraph==8.0.0
    # via azure-haymaker (pyproject.toml)
azure-mgmt-subscription==3.1.1
    # via azure-haymaker (pyproject.toml)
azure-monitor-ingestion==1.1.0
    # via azure-haymaker (pyproject.toml)
azure-servicebus==7.14.3
    # via azure-haymaker (pyproject.toml)
azure-storage-blob==12.27.1
    # via azure-haymaker (pyproject.toml)
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   msrest
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
cryptography==46.0.3
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
distro==1.9.0
    # via anthropic
docstring-parser==0.17.0
    # via anthropic
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
furl==2.1.4
    # via azure-functions-durable
h11==0.16.0
    # via httpcore
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   azure-haymaker (pyproject.toml)
    #   anthropic
    #   microsoft-kiota-http
    #   msgraph-core
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via opentelemetry-api
isodate==0.7.2
    # via
    #   azure-data-tables
    #   azure-keyvault-secrets
    #   azure-mgmt-authorization
    #   azure-mgmt-containerinstance
    #   azure-mgmt-resource
    #   azure-monitor-ingestion
    #   azure-servicebus
    #   azure-storage-blob
    #   msrest
jiter==0.12.0
    # via anthropic
markupsafe==3.0.3
    # via werkzeug
microsoft-kiota-abstractions==1.9.7
    # via
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   microsoft-kiota-serialization-form
    #   microsoft-kiota-serialization-json
    #   microsoft-kiota-serialization-multipart
    #   microsoft-kiota-serialization-text
    #   msgraph-core
microsoft-kiota-authentication-azure==1.9.7
    # via msgraph-core
microsoft-kiota-http==1.9.7
    # via msgraph-core
microsoft-kiota-serialization-form==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-json==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-multipart==1.9.7
    # via msgraph-sdk
microsoft-kiota-serialization-text==1.9.7
    # via msgraph-sdk
msal==1.34.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
msgraph-core==1.3.8
    # via msgraph-sdk
msgraph-sdk==1.48.0
    # via azure-haymaker (pyproject.toml)
msrest==0.7.1
    # via
    #   azure-mgmt-resourcegraph
    #   azure-mgmt-subscription
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
oauthlib==3.3.1
    # via requests-oauthlib
opentelemetry-api==1.38.0
    # via
    #   azure-functions-durable
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
opentelemetry-sdk==1.38.0
    # via
    #   azure-functions-durable
    #   microsoft-kiota-abstractions
    #   microsoft-kiota-authentication-azure
    #   microsoft-kiota-http
opentelemetry-semantic-conventions==0.59b0
    # via opentelemetry-sdk
orderedmultidict==1.0.1
    # via furl
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pycparser==2.23
    # via cffi
pydantic==2.12.4
    # via
    #   azure-haymaker (pyproject.toml)
    #   anthropic
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via azure-haymaker (pyproject.toml)
pyjwt==2.10.1
    # via msal
python-dateutil==2.9.0.post0
    # via azure-functions-durable
python-dotenv==1.2.1
    # via
    #   azure-haymaker (pyproject.toml)
    #   pydantic-settings
requests==2.32.5
    # via
    #   azure-core
    #   azure-functions-durable
    #   msal
    #   msrest
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via msrest
six==1.17.0
    # via
    #   furl
    #   orderedmultidict
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anthropic
    #   anyio
std-uritemplate==2.0.8
    # via microsoft-kiota-abstractions
typing-extensions==4.15.0
    # via
    #   anthropic
    #   azure-core
    #   azure-cosmos
    #   azure-data-tables
    #   azure-identity
    #   azure-keyvault-secrets
    #   azure-mgmt-resource
    #   azure-monitor-ingestion
    #   azure-servicebus
    #   azure-storage-blob
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via azure-functions
yarl==1.22.0
    # via
    #   aiohttp
    #   azure-data-tables
zipp==3.23.0
    # via importlib-metadata
